+++ src/App/Helpers/ConfigGlobal.php
@@ -18,24 +18,49 @@ class ConfigGlobal
     static $basePath = 'Helpers\DataFinder\\';
     static $file_not_exist_message = 'Configuration file for the package not found. Please make sure you have correct configuration file setup.';
 
+    // Configuration File Section Constants
+    const MAIN_CONFIG_FILE = ['key' => 'main_config_file', 'stub_path' => '/../stubs/main.stub'];
+    const SECTION_A = ['key' => 'frontend_base', 'stub_path' => '/../stubs/frontend/base.stub'];
+    const SECTION_B = ['key' => 'backend_primary', 'stub_path' => '/../stubs/database/primary.stub'];
+    const SECTION_C = ['key' => 'frontend_filters', 'stub_path' => '/../stubs/frontend/filters.stub'];
+    const SECTION_D = ['key' => 'backend_joins', 'stub_path' => '/../stubs/database/joins.stub'];
+    const SECTION_E = ['key' => 'frontend_table_headers', 'stub_path' => '/../stubs/frontend/table_headers.stub'];
+    const SECTION_F = ['key' => 'frontend_row_buttons', 'stub_path' => '/../stubs/frontend/row_buttons.stub'];
 
     public static function validateConfigFile($path)
     {
-        if(file_exists(app_path(self::$basePath . $path . '.php'))){
+        if(file_exists(self::getPath($path))){
             return true;
         }
         return false;
     }
-
 
     public static function getValueFromFile($path, $value_to_get)
     {
         // This function is used to get the array value from the file contains configrations
         // only as array. Not a class but just a file.
 
-        $file = include(app_path(self::$basePath . $path . '.php'));
+        $file = include(self::getPath($path));
         $value = Arr::get($file, $value_to_get);
         return $value;
     }
 
+    public static function getPath($path)
+    {
+        $file_path = app_path(self::$basePath . $path . '.php');
+        $path_info = pathinfo($file_path);
+        self::createDirectoryIfNotExist($path_info['dirname']);
+        return $file_path;
+    }
+
+    public static function createDirectoryIfNotExist($path, $replace = false)
+    {
+        if (file_exists($path) && $replace) {
+            rmdir($path);
+        }
+
+        if (!file_exists($path)) {
+            mkdir($path, 0755, true);
+        }
+    }
 }